token authorization for all post and get method

+++ src/controller/AccountDeleteById.js
@@ -10,22 +10,36 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 Object.defineProperty(exports, "__esModule", { value: true });
 const typeorm_1 = require("typeorm");
 const Account_1 = require("../entity/Account");
+const User = require("../controller/UserController");
 
 function accountDeleteByIdAction(request, response) {
-    return __awaiter(this, void 0, void 0, function* () {
+    let authorizationHeader = request.headers['authorization'] || request.headers['Authorization']
+    if (typeof authorizationHeader !== 'undefined') {
+    let [, token] = authorizationHeader.split(' ');
     
-        const accountRepository = typeorm_1.getManager().getRepository(Account_1.Account);
-     
-        const account = yield accountRepository.delete(request.params.id);
+    if (token != User.getToken()) {
+        response.sendStatus(403) // Forbidden, you're not logged in
+        console.log("User not logged in");
+    } else {
+        return __awaiter(this, void 0, void 0, function* () {
+        
+            const accountRepository = typeorm_1.getManager().getRepository(Account_1.Account);
+        
+            const account = yield accountRepository.delete(request.params.id);
 
-        if (!account) {
-            response.status(404);
-            response.end();
-            return;
-        }
-   
-        response.send(account);
-    });
+            if (!account) {
+                response.status(404);
+                response.end();
+                return;
+            }
+    
+            response.send(account);
+        });
+    }
+     
+  } else {
+             response.sendStatus(403);
+    }
 }
 exports.accountDeleteByIdAction = accountDeleteByIdAction;
 //# sourceMappingURL=AccountGetByIdAction.js.map
